move metrics to tools/
